requirements.txt
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
streamlit==1.20.0
Pillow==9.4.0
requests==2.28.2
azure-storage-blob==12.14.1
